Remove default value parameter from Parser.GetAuthor
The point of this method was just to make a simpler call to a common TOC
field. Having a default value parameter that can be used if the Author field
is missing kind of defeats that point as it would be just as easy to call
GetEntry('Author', 'John Doe').